+++ crates/runtime/src/lib.rs
@@ -24,9 +24,11 @@ use cairo_vm::vm::errors::vm_errors::VirtualMachineError;
 use cairo_vm::vm::runners::cairo_runner::{ResourceTracker, RunResources};
 use cairo_vm::vm::vm_core::VirtualMachine;
 use conversions::byte_array::ByteArray;
+use conversions::serde::deserialize::BufferReadError;
 use conversions::serde::deserialize::BufferReader;
 use conversions::serde::serialize::raw::RawFeltVec;
 use conversions::serde::serialize::{CairoSerialize, SerializeToFeltVec};
+use indoc::indoc;
 use starknet_api::StarknetApiError;
 use std::any::Any;
 use std::collections::HashMap;
@@ -34,7 +36,6 @@ use std::io;
 use thiserror::Error;
 
 pub mod starknet;
-pub mod utils;
 
 // from core/src/starknet/testing.cairo
 const CAIRO_TEST_CHEATCODES: [&str; 14] = [
@@ -102,7 +103,7 @@ impl<'a> SignalPropagator for StarknetRuntime<'a> {
     fn propagate_cheatcode_signal(&mut self, _selector: &str, _inputs: &[Felt252]) {}
 }
 
-pub fn parse_selector(selector: &BigIntAsHex) -> Result<String, HintError> {
+fn parse_selector(selector: &BigIntAsHex) -> Result<String, HintError> {
     let selector = &selector.value.to_bytes_be().1;
     let selector = std::str::from_utf8(selector)
         .map_err(|_| CustomHint(Box::from("Failed to parse the cheatcode selector")))?;
@@ -185,16 +186,16 @@ impl<Extension: ExtensionLogic> HintProcessorLogic for ExtendedRuntime<Extension
         constants: &HashMap<String, Felt252>,
     ) -> Result<(), HintError> {
         let maybe_extended_hint = hint_data.downcast_ref::<Hint>();
-        if let Some(Hint::Starknet(starknet_hint)) = maybe_extended_hint {
-            if let StarknetHint::Cheatcode {
-                selector,
-                input_start,
-                input_end,
-                output_start,
-                output_end,
-            } = starknet_hint
-            {
-                return self.execute_cheatcode_hint(
+
+        match maybe_extended_hint {
+            Some(Hint::Starknet(starknet_hint)) => match starknet_hint {
+                StarknetHint::Cheatcode {
+                    selector,
+                    input_start,
+                    input_end,
+                    output_start,
+                    output_end,
+                } => self.execute_cheatcode_hint(
                     vm,
                     exec_scopes,
                     hint_data,
@@ -206,14 +207,15 @@ impl<Extension: ExtensionLogic> HintProcessorLogic for ExtendedRuntime<Extension
                         output_start,
                         output_end,
                     },
-                );
-            } else if let StarknetHint::SystemCall { system } = starknet_hint {
-                return self.execute_syscall_hint(vm, exec_scopes, hint_data, constants, system);
-            }
+                ),
+                StarknetHint::SystemCall { system } => {
+                    self.execute_syscall_hint(vm, exec_scopes, hint_data, constants, system)
+                }
+            },
+            _ => self
+                .extended_runtime
+                .execute_hint(vm, exec_scopes, hint_data, constants),
         }
-
-        self.extended_runtime
-            .execute_hint(vm, exec_scopes, hint_data, constants)
     }
 
     fn compile_hint(
@@ -465,4 +467,18 @@ pub enum EnhancedHintError {
     StarknetApi(#[from] StarknetApiError),
     #[error("Failed to parse {path} file")]
     FileParsing { path: String },
+}
+
+impl From<BufferReadError> for EnhancedHintError {
+    fn from(value: BufferReadError) -> Self {
+        EnhancedHintError::Anyhow(
+            anyhow::Error::from(value)
+                .context(
+                    indoc!(r"
+                        Reading from buffer failed, this can be caused by calling starknet::testing::cheatcode with invalid arguments.
+                        Probably `snforge_std`/`sncast_std` version is incompatible, check above for incompatibility warning.
+                    ")
+                )
+        )
+    }
 }
